Debut algo

+++ include/game.h
@@ -6,7 +6,7 @@
 /*   By: fhuang <fhuang@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/02/08 12:10:39 by fhuang            #+#    #+#             */
-/*   Updated: 2017/02/09 17:33:35 by fhuang           ###   ########.fr       */
+/*   Updated: 2017/03/25 17:57:40 by fhuang           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -25,12 +25,14 @@
 #  define ERROR_STRING "Error\n"
 # endif
 
+# include <inttypes.h>
+
 typedef struct	s_piles
 {
-	int			len;
-	int			len_a;
+	uint16_t	len;
+	uint16_t	len_a;
 	int			*pile_a;
-	int			len_b;
+	uint16_t	len_b;
 	int			*pile_b;
 	int			*sorted_pile;
 }				t_piles;
@@ -44,20 +46,20 @@ typedef struct	s_instruction1
 typedef struct	s_instruction2
 {
 	const char	*str;
-	void		(*f)(int **pile, int len);
+	void		(*f)(int **pile, uint16_t len);
 }				t_instruction2;
 
 typedef struct	s_instruction3
 {
 	const char	*str;
-	void		(*f)(int **src, int **dst, int *len_src, int *len_dst);
+	void		(*f)(int **src, int **dst, uint16_t *len_src, uint16_t *len_dst);
 }				t_instruction3;
 
-char**			check_parameters(char **av, int *len);
+char**			check_parameters(char **av, uint16_t *len);
 
 int				call_instruction(t_piles *piles, const char *str);
 
-int				is_pile_sorted(int *pile, int len);
+int				is_pile_sorted(int *pile, uint16_t len);
 int				is_game_finished(t_piles piles);
 
 void			init_piles(t_piles *piles, char **av);
@@ -67,10 +69,10 @@ void			destroy_piles(t_piles *piles);
 **	INSTRUCTIONS
 */
 
-void			push_to(int **src, int **dst, int *len_src, int *len_dst);
-void			swap(int **pile, int len);
-void			rotate(int **pile, int len);
-void			reverse_rotate(int **pile, int len);
+void			push_to(int **src, int **dst, uint16_t *len_src, uint16_t *len_dst);
+void			swap(int **pile, uint16_t len);
+void			rotate(int **pile, uint16_t len);
+void			reverse_rotate(int **pile, uint16_t len);
 void			swap_both_piles(t_piles *piles);
 void			rotate_both_piles(t_piles *piles);
 void			reverse_rotate_both_piles(t_piles *piles);
