+++ src/instructions/reverse_rotate.c
@@ -6,11 +6,13 @@
 /*   By: fhuang <fhuang@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/02/08 18:53:34 by fhuang            #+#    #+#             */
-/*   Updated: 2017/02/09 13:40:11 by fhuang           ###   ########.fr       */
+/*   Updated: 2017/03/25 17:57:09 by fhuang           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-void	reverse_rotate(int **pile, int len)
+#include <inttypes.h>
+
+void	reverse_rotate(int **pile, uint16_t len)
 {
 	int		tmp;
 	int		i;
