+++ src/instructions/rotate.c
@@ -6,11 +6,13 @@
 /*   By: fhuang <fhuang@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/02/08 18:44:06 by fhuang            #+#    #+#             */
-/*   Updated: 2017/02/09 13:33:07 by fhuang           ###   ########.fr       */
+/*   Updated: 2017/03/25 17:57:18 by fhuang           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-void	rotate(int **pile, int len)
+#include <inttypes.h>
+
+void	rotate(int **pile, uint16_t len)
 {
 	int		tmp;
 	int		i;
