+++ src/instructions/swap.c
@@ -6,13 +6,13 @@
 /*   By: fhuang <fhuang@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/02/08 18:50:55 by fhuang            #+#    #+#             */
-/*   Updated: 2017/02/09 13:10:41 by fhuang           ###   ########.fr       */
+/*   Updated: 2017/03/25 17:56:55 by fhuang           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "game.h"
 
-void	swap(int **pile, int len)
+void	swap(int **pile, uint16_t len)
 {
 	if (!pile || len < 2)
 		return ;
